[FIXED] Issues with currency tab and player stucking in world, with improved patch from http://arcemu.org/forums/index.php?showtopic=17169 . Report any bugs on forums.

+++ src/arcemu-world/Container.cpp
@@ -129,7 +129,7 @@ bool Container::HasItems()
 	return false;
 }
 
-bool Container::AddItem(int8 slot, Item *item)
+bool Container::AddItem(int16 slot, Item *item)
 {
 	if (slot < 0 || (uint32)slot >= GetProto()->ContainerSlots)
 		return false;
@@ -235,7 +235,7 @@ void Container::SwapItems(int8 SrcSlot, int8 DstSlot)
 	}
 }
 
-Item *Container::SafeRemoveAndRetreiveItemFromSlot(int8 slot, bool destroy)
+Item *Container::SafeRemoveAndRetreiveItemFromSlot(int16 slot, bool destroy)
 {
 	if (slot < 0 || (uint32)slot >= GetProto()->ContainerSlots)
 		return false;
@@ -265,7 +265,7 @@ Item *Container::SafeRemoveAndRetreiveItemFromSlot(int8 slot, bool destroy)
 	return pItem;
 }
 
-bool Container::SafeFullRemoveItemFromSlot(int8 slot)
+bool Container::SafeFullRemoveItemFromSlot(int16 slot)
 {
 	if (slot < 0 || (uint32)slot >= GetProto()->ContainerSlots)
 		return false;
