+++ src/arcemu-world/Container.h
@@ -31,11 +31,11 @@ public:
 	void Create( uint32 itemid, Player *owner );
 	void LoadFromDB( Field*fields);
 
-	bool AddItem(int8 slot, Item *item);
+	bool AddItem(int16 slot, Item *item);
 	bool AddItemToFreeSlot(Item *pItem, uint32 * r_slot);
-	Item *GetItem(int8 slot)
+	Item *GetItem(int16 slot)
 	{
-		if(slot >= 0 && (uint8)slot < GetProto()->ContainerSlots)
+		if(slot >= 0 && (uint16)slot < GetProto()->ContainerSlots)
 			return m_Slot[slot];
 		else
 			return 0;
@@ -45,8 +45,8 @@ public:
 	bool HasItems();
 	
 	void SwapItems(int8 SrcSlot,int8 DstSlot);
-	Item *SafeRemoveAndRetreiveItemFromSlot(int8 slot, bool destroy); //doesn't destroy item from memory
-	bool SafeFullRemoveItemFromSlot(int8 slot); //destroys item fully
+	Item *SafeRemoveAndRetreiveItemFromSlot(int16 slot, bool destroy); //doesn't destroy item from memory
+	bool SafeFullRemoveItemFromSlot(int16 slot); //destroys item fully
    
 	void SaveBagToDB(int8 slot, bool first, QueryBuffer * buf);
